Upgrade to 1.5.0 (read the commit)
- Upgrade from Laravel 10 to 11
- Dropped PHP 8.1
- Seems laravel dropped APC cache
- Added some customizations in .env
- The default session driver is now "database"; it obviously requires running db migrations. If you are upgrading you have probably "SESSION_DRIVER" set to "file" in your .env. You have to change it manually to "database"
- Cache driver is kept to "file" as default. You may change it to "database", but usually providers do a huge user compression in databases, so it may slow down your whole website. Usually local file system (with SSD) is better, but you may need to test
- The default mailer is local sendmail
- Added some tables to handle cache and sessions in the database. Please remember to run `php artisan migrate` to update the schema if you want to try them

+++ database/migrations/2020_08_31_145624_create_comics_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
     public function up() {
         Schema::create('comics', function (Blueprint $table) {
             $table->id();
-            $table->float('order_index', 8, 3);
+            $table->decimal('order_index', 8, 3);
             $table->string('name');
             $table->string('slug')->unique();
             $table->string('salt');
